add phoneSaved popup to Header
