+++ src/components/shared/popups/styles/PhoneSavedPopup.module.css
@@ -9,12 +9,10 @@
     color: var(--text-color-1);
     opacity: 0;
     position: absolute;
-    /* margin-left: 453px; */
     top: 56px;
-    right: 17%;
+    left: 50%;
     visibility: hidden;
     z-index: 3;
-            
     transition: all 0.4s ease-in-out;
 }
 
@@ -30,17 +28,11 @@
     width: 100%;
     gap: 12px;
     margin-left: 20px;
-    /* margin-right: 15px; */
-}
-
-.imgCloseContainer {
-    /* margin-left: 16px; */
 }
 
 .imgClose{
     position: relative;
     top: -22px;
-    /* right: -18px; */
     cursor: pointer;
 }
